Extract concentration analytics router
