+++ tests/support/app_runtime.py
@@ -4,6 +4,7 @@ from contextlib import contextmanager
 from typing import Any, Iterator
 
 import app.main as main_module
+import app.routers.concentration as concentration_module
 import app.routers.drawdown as drawdown_module
 import app.routers.risk_calculation as risk_calculation_module
 import app.routers.rolling as rolling_module
@@ -32,6 +33,7 @@ def override_app_runtime(
     )
     original_rolling_performance_class: Any = getattr(rolling_module, "LotusPerformanceClient")
     original_core_class: Any = getattr(main_module, "LotusCoreClient")
+    original_concentration_core_class: Any = getattr(concentration_module, "LotusCoreClient")
     original_rolling_core_class: Any = getattr(rolling_module, "LotusCoreClient")
 
     try:
@@ -50,6 +52,7 @@ def override_app_runtime(
             setattr(rolling_module, "LotusPerformanceClient", lotus_performance_class)
         if lotus_core_class is not _UNSET:
             setattr(main_module, "LotusCoreClient", lotus_core_class)
+            setattr(concentration_module, "LotusCoreClient", lotus_core_class)
             setattr(rolling_module, "LotusCoreClient", lotus_core_class)
         if dependency_statuses is not _UNSET:
             app.state.dependency_statuses = dependency_statuses
@@ -67,4 +70,5 @@ def override_app_runtime(
         )
         setattr(rolling_module, "LotusPerformanceClient", original_rolling_performance_class)
         setattr(main_module, "LotusCoreClient", original_core_class)
+        setattr(concentration_module, "LotusCoreClient", original_concentration_core_class)
         setattr(rolling_module, "LotusCoreClient", original_rolling_core_class)
